Added buffer abstraction, clang-tidy, and updated serial class. buffer: Added buffer.hpp which abstracts pointers passed to read/write functions (from std::experimental::net). clang-tidy: Added clang-tidy support. serial: Added more read functions, small fixes and formatting. serial: Updated to use new buffer abstractions instead of raw pointers and std::vector<unsigned char>.

+++ src/periphery/i2c.cpp
@@ -110,7 +110,7 @@ void I2C::transfer_impl(uint16_t addr, ForwardIt first, ForwardIt last) const
         auto& m = (Message&) *first++;
         p[i].addr  = addr;
         p[i].flags = m.flags;
-        p[i].len   = m.data.size();
+        p[i].len   = (uint16_t) m.data.size();
         p[i].buf   = m.data.data();
     }
